requirement.txt
pandas
mlflow
notebook
numpy 
scikit-learn
matplotlib 
python-box
pyYAML
tqdm
ensure 
joblib
types-pyYAML
Flask
Flask-Cors